Correct autocomplete when managing users
MAJOR: Fixed a problem with how autocomplete is handled in the presence
of the shibboleth module, such that adding a user to a group creates a
duplicate user instead of adding the user appropriately. This is because
the shibboleth module changes the field used as the unique identifier
from Email to UniqueIdentifier.